third_party/WebKit/LayoutTests/fast/dom/custom/element-upgrade.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id="container"></div>
<script>
description('Tests the element upgrade algorithm.');

// "Element Upgrade" is the processing of custom elements which were
// created before their definition was available, when the definition
// becomes available. The following scenarios cover a lot but are not
// exhaustive.

// Scenario A: Custom tag; upgrade candidate is not in the document;
// upgrade candidate did not have a JavaScript wrapper at upgrade
// time; custom element does not have a created callback.
var host = document.createElement('div');
host.innerHTML = '<x-a></x-a>';  // Using innerHTML avoids wrapping x-a
var A = document.registerElement('x-a', {prototype: Object.create(HTMLElement.prototype)});
shouldBeTrue('host.firstChild instanceof A');

// Scenario B: Type extension; upgrade candidate is in the document;
// upgrade candidate did have a JavaScript wrapper at upgrade time;
// custom element has a created callback.
var element = document.createElement('span', 'x-b');
var proto = Object.create(HTMLSpanElement.prototype);
var callCount = 0;
proto.createdCallback = function () {
  callCount++;
};
var B = document.registerElement('x-b', {extends: 'span', prototype: proto});
shouldBeTrue('element instanceof B');
shouldBe('callCount', '1');

// Scenario C: The candidate is a custom tag but the definition is a
// type extension. Upgrade should not happen.
element = document.createElement('x-c');
var C = document.registerElement('x-c', {extends: 'span', prototype: Object.create(HTMLSpanElement.prototype)});
shouldBeFalse('element instanceof C');
shouldBe('Object.getPrototypeOf(element)', 'HTMLElement.prototype');

// Scenario D: The candidate is a type extension, but the definition
// extends a different tag. Upgrade should not happen.
document.body.appendChild(host);
host.innerHTML = '<span is="x-d"></span>';
var D = document.registerElement('x-d', {extends: 'div', prototype: Object.create(HTMLDivElement.prototype)});
shouldBeFalse('host.firstChild instanceof D');
shouldBe('document.querySelector(":unresolved")', 'host.firstChild');

// Scenario E: The order of upgrades should be the order of completing parsing.
// Use a good number of elements to avoid false positives from random correct ordering.
host.innerHTML = '<x-e id="e1"><x-e id="e2"></x-e></x-e><x-e id="e3"></x-e><x-e id="e4"></x-e><x-e id="e5"></x-e>';
var upgradedOrder = [];
var protoE = Object.create(HTMLElement.prototype);
protoE.createdCallback = function() { upgradedOrder.push(this.id); };
document.registerElement('x-e', {prototype: protoE});
shouldBe('upgradedOrder', '["e1","e2","e3","e4","e5"]');

successfullyParsed = true;
</script>
